Split navigation data to separate file
