+++ src/components/Navigation/data/navigationData.ts
@@ -1,0 +1,13 @@
+interface NavLinkProps {
+  name: string;
+  source: string;
+}
+
+const navLinkData: NavLinkProps[] = [
+  { name: "Home", source: "home" },
+  { name: "About", source: "about" },
+  { name: "Services", source: "services" },
+  { name: "Testimonials", source: "testimonials" },
+];
+
+export default navLinkData;
